Add prompts to create preferences file and add/update .gitignore

+++ src/_write/prefs.js
@@ -2,24 +2,28 @@ let { parser, lexer, stringify } = require('@architect/parser')
 let fs = require('fs')
 let { existsSync, readFileSync } = fs
 let { join } = require('path')
+let waterfall = require('run-waterfall')
+let prompt = require('./prompt')
 
-module.exports = function write ({ envVars, update }) {
+module.exports = function write ({ envVars, update }, callback) {
+  let cwd = process.cwd()
   // Establish filename, uh, preference
   let short = 'prefs.arc'
   let long = 'preferences.arc'
   let filename
   let prefsPath
-  if (existsSync(join(process.cwd(), short))) {
+  if (existsSync(join(cwd, short))) {
     filename = short
-    prefsPath = join(process.cwd(), short)
+    prefsPath = join(cwd, short)
   }
   else {
     filename = long
-    prefsPath = join(process.cwd(), long)
+    prefsPath = join(cwd, long)
   }
 
   let arc
-  if (existsSync(prefsPath)) {
+  let prefsExists = existsSync(prefsPath)
+  if (prefsExists) {
     let raw = readFileSync(prefsPath).toString()
     arc = parser(lexer(raw))
     delete arc.env
@@ -64,9 +68,60 @@ ${staging}
 ${production}
 `
 
-  let prefs = stringify(arc)
-  prefs = prefs ? prefs + '\n' : prefs
-  let contents = prefs + envString
-  fs.writeFileSync(prefsPath, contents)
-  update.done(`Updated ${filename} file with latest environment variables`)
+  let gitignore
+  let gitignorePath = join(cwd, '.gitignore')
+  let hasGitignore = existsSync(gitignorePath)
+  waterfall([
+    function (callback) {
+      if (prefsExists) callback(null, true)
+      else {
+        let question = 'Would you like to create a local preferences file? [Y|n]'
+        let why = 'Sandbox uses this to load your environment variables'
+        update.status(question, why)
+        prompt(update, callback)
+      }
+    },
+    function (write, callback) {
+      if (write) {
+        let prefs = stringify(arc)
+        prefs = prefs ? prefs + '\n' : prefs
+        let contents = prefs + envString
+        fs.writeFileSync(prefsPath, contents)
+        update.done(`Updated ${filename} file with latest environment variables`)
+        callback(null, true)
+      }
+      else callback(null, false)
+    },
+    function (wrote, callback) {
+      let isGitRepo = existsSync(join(cwd, '.git'))
+      let question = 'Would you like Git to ignore your preferences file (strongly recommended)? [Y|n]'
+      if (isGitRepo && (prefsExists || wrote)) {
+        if (!hasGitignore) {
+          gitignore = filename + '\n'
+          update.status(question)
+          prompt(update, callback)
+        }
+        else {
+          gitignore = readFileSync(gitignorePath).toString()
+          let isIgnored = gitignore.split('\n').some(l => l.includes(filename))
+          if (isIgnored) callback(null, false)
+          else {
+            gitignore = gitignore.trim() + `\n${filename}\n`
+            update.status(question)
+            prompt(update, callback)
+          }
+        }
+      }
+      else callback(null, false)
+    },
+    function (write, callback) {
+      if (write) {
+        fs.writeFileSync(gitignorePath, gitignore)
+        let action = hasGitignore ? 'Updated' : 'Created'
+        update.done(`${action} .gitignore with ${filename}`)
+        callback()
+      }
+      else callback()
+    }
+  ], callback)
 }
